sink AI检视问题修复

+++ services/codec/src/video_sink_decoder.cpp
@@ -241,7 +241,7 @@ bool VideoSinkDecoder::DecodeVideoData(const char *data, int32_t size, uint64_t 
     }
 
     MediaAVCodec::AVCodecBufferInfo bufferInfo;
-    bufferInfo.presentationTimeUs = pts;
+    bufferInfo.presentationTimeUs = static_cast<int64_t>(pts);
     bufferInfo.size = size;
     bufferInfo.offset = 0;
     WfdSinkHiSysEvent::GetInstance().RecordMediaDecodeStartTime(MediaReportType::VIDEO, bufferInfo.presentationTimeUs);
